+++ src/MainWindow/ArgsPrice.ts
@@ -1,0 +1,11 @@
+class ArgsPrice{
+    ride: number;
+    price: number;
+    isPrimaryPrice: boolean;
+
+    constructor(ride: number, price: number, isPrimaryPrice: boolean){
+        this.ride = ride;
+        this.price = price;
+        this.isPrimaryPrice = isPrimaryPrice;
+    }
+}