+++ src/MainWindow/ArgsSetting.ts
@@ -1,0 +1,11 @@
+class ArgsSetting{
+    ride: Ride;
+    setting: number;
+    value: number;
+
+    constructor(ride: Ride, setting: number, value: number){
+        this.ride = ride;
+        this.setting = setting;
+        this.value = value;
+    }
+}