+++ src/RideWindow/ViewportArray.ts
@@ -3,8 +3,17 @@ export class ViewportArray {
 
     stationName: string
 
+    vehicleNumber: number = -100000000000000000
     constructor(coordsXYZ: CoordsXYZ, stationName: string) {
         this.coordsXYZ = coordsXYZ
         this.stationName = stationName
+    }
+
+    setVehicleNumber(vehicleNumber: number) {
+        this.vehicleNumber = vehicleNumber
+    }
+
+    getVehicleNumber() {
+        return this.vehicleNumber
     }
 }